# contracts/identity.openapi.yaml
openapi: 3.0.3
info:
  title: Zennyt - Identity API
  description: >
    Identity bounded context: local, Google, and Apple authentication, JWT lifecycle,
    candidate/student and recruiter onboarding, and professional profiles.
  version: 2.0.0
servers:
  - url: https://api.zennyt.com/api/v1
    description: Production
  - url: https://api-staging.zennyt.com/api/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /onboarding/candidate-student:
    post:
      operationId: postOnboardingCandidateStudent
      tags: [Onboarding]
      summary: Create candidate or student onboarding information
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CandidateStudentOnboardingInput' }
      responses:
        '201':
          description: Onboarding created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CandidateStudentOnboarding' }
        '403': { $ref: './common.openapi.yaml#/components/responses/Forbidden' }
        '409': { $ref: './common.openapi.yaml#/components/responses/Conflict' }
  /onboarding/candidate-student/me:
    get:
      operationId: getOnboardingCandidateStudentMe
      tags: [Onboarding]
      summary: Get current candidate or student onboarding information
      responses:
        '200':
          description: Current onboarding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CandidateStudentOnboarding' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
    put:
      operationId: putOnboardingCandidateStudentMe
      tags: [Onboarding]
      summary: Update current candidate or student onboarding information
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CandidateStudentOnboardingInput' }
      responses:
        '200':
          description: Onboarding updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CandidateStudentOnboarding' }
  /onboarding/recruiter:
    post:
      operationId: postOnboardingRecruiter
      tags: [Onboarding]
      summary: Create recruiter onboarding information
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecruiterOnboardingInput' }
      responses:
        '201':
          description: Onboarding created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecruiterOnboarding' }
        '403': { $ref: './common.openapi.yaml#/components/responses/Forbidden' }
        '409': { $ref: './common.openapi.yaml#/components/responses/Conflict' }
  /onboarding/recruiter/me:
    get:
      operationId: getOnboardingRecruiterMe
      tags: [Onboarding]
      summary: Get current recruiter onboarding information
      responses:
        '200':
          description: Current onboarding
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecruiterOnboarding' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
    put:
      operationId: putOnboardingRecruiterMe
      tags: [Onboarding]
      summary: Update current recruiter onboarding information
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecruiterOnboardingInput' }
      responses:
        '200':
          description: Onboarding updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecruiterOnboarding' }
  /onboarding/recruiter/me/logo:
    post:
      operationId: postOnboardingRecruiterMeLogo
      tags: [Onboarding]
      summary: Upload (or replace) the company logo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (PNG, JPEG or WEBP, max 5MB).
      responses:
        '200':
          description: Logo uploaded; onboarding with companyLogoUrl set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecruiterOnboarding' }
        '400': { $ref: './common.openapi.yaml#/components/responses/BadRequest' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
    delete:
      operationId: deleteOnboardingRecruiterMeLogo
      tags: [Onboarding]
      summary: Remove the company logo
      responses:
        '200':
          description: Logo removed; onboarding with companyLogoUrl cleared
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RecruiterOnboarding' }
        '404': { $ref: './common.openapi.yaml#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { $ref: './common.openapi.yaml#/components/securitySchemes/bearerAuth' }
...
  schemas:
...
    CandidateStudentOnboardingInput:
      type: object
      properties:
        school: { type: string, maxLength: 150 }
        educationLevel: { type: string, maxLength: 150 }
        fieldOfWork: { type: string, maxLength: 150 }
        lastPositionHeld: { type: string, maxLength: 150 }
        yearsOfExperience: { type: integer, minimum: 0 }
        cvFileUrl: { type: string, maxLength: 500 }
    CandidateStudentOnboarding:
      allOf:
        - $ref: '#/components/schemas/CandidateStudentOnboardingInput'
        - $ref: '#/components/schemas/OwnedResource'
    RecruiterOnboardingInput:
      type: object
      required:
        [jobTitle, companyName, companySize, fieldOfWork, companyLocation, companyRegistrationNumber]
      properties:
        jobTitle: { type: string, maxLength: 150 }
        companyName: { type: string, maxLength: 150 }
        companySize: { type: string, maxLength: 100 }
        fieldOfWork: { type: string, maxLength: 150 }
        companyLocation: { type: string, maxLength: 150 }
        companyRegistrationNumber: { type: string, maxLength: 100 }
        aboutMe: { type: string, maxLength: 1000 }
    RecruiterOnboarding:
      allOf:
        - $ref: '#/components/schemas/RecruiterOnboardingInput'
        - $ref: '#/components/schemas/OwnedResource'
        - type: object
          properties:
            # Le logo est en lecture seule : il se gère via POST/DELETE /onboarding/recruiter/me/logo.
            companyLogoUrl: { type: string, maxLength: 500, readOnly: true }
...
    OwnedResource:
      type: object
      required: [id, userId, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...